+++ spec/support/flush_default_ractor_warning.rb
@@ -1,0 +1,4 @@
+previous_verbose = $VERBOSE
+$VERBOSE = nil
+defined?(Ractor) && Ractor.new {}
+$VERBOSE = previous_verbose